requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2023.11.17
python-telegram-bot==13.7
pytz==2023.3.post1
six==1.16.0
tornado==6.3.3
tzdata==2023.3
tzlocal==5.2
flake8==3.9.2
flake8-docstrings==1.6.0
pytest==6.2.5
pytest-timeout==2.1.0
python-dotenv==0.19.0
requests==2.26.0
